Move item interaction code into unified actions
